Move Directory of Privacy
