+++ src/Components/Privacy/Privacy.css
